fix(form & root): move the useQuery files and api files from form module to root of the core package
