+++ packages/core/src/modules/form/src/components/actions/submit/submit.loading.tsx
@@ -1,6 +1,6 @@
 import { Skeleton } from '@mui/material';
 
-import { LoadingProps } from '../../../types/configs.type';
+import { LoadingProps } from '../../../../../../types/configs.type';
 import useSubmitLoading from './useSubmit.loading';
 
 const SubmitLoading = (props: LoadingProps) => {
