+++ packages/core/src/modules/form/src/components/actions/submit/useSubmit.loading.ts
@@ -1,4 +1,4 @@
-import { LoadingProps } from '../../../types/configs.type';
+import { LoadingProps } from '../../../../../../types/configs.type';
 
 const useSubmitLoading = (props: LoadingProps) => {
   const { animation = 'wave', sx, ...otherProps } = props;
