+++ packages/core/src/modules/form/src/components/builder/formBuilder.types.ts
@@ -1,6 +1,6 @@
 import { FieldProps, FormTypes } from '../../types/public.types';
 
-import { Configs } from '../../types/configs.type';
+import { Configs } from '../../../../../types/configs.type';
 
 export type FormBuilderProps = {
   fieldType: FormTypes;
