+++ packages/core/src/modules/form/src/components/fields/text/text.loading.tsx
@@ -1,6 +1,6 @@
 import { Skeleton } from '@mui/material';
 
-import { LoadingProps } from '../../../types/configs.type';
+import { LoadingProps } from '../../../../../../types/configs.type';
 import useTextLoading from './useText.loading';
 
 const TextLoading = (props: LoadingProps) => {
