+++ packages/core/src/modules/form/src/components/fields/text/text.types.ts
@@ -1,6 +1,6 @@
 import { TextFieldProps } from '@mui/material';
 
-import { Api } from '../../../types/api.types';
+import { Api } from '../../../../../../types/api.types';
 import { Dependesies, FormId, Id } from '../../../types/public.types';
 import { Script } from '../../../types/script.types';
 import { Rule } from '../../../types/validation.types';
