+++ packages/core/src/modules/form/src/components/fields/text/useText.loading.ts
@@ -1,4 +1,4 @@
-import { LoadingProps } from '../../../types/configs.type';
+import { LoadingProps } from '../../../../../../types/configs.type';
 
 const useTextLoading = (props: LoadingProps) => {
   const { animation = 'wave', sx, ...otherProps } = props;
