+++ packages/core/src/modules/form/src/index.ts
@@ -1,6 +1,6 @@
 import FormBuilder from './components/builder/formBuilder';
 
 export type { FieldProps, FormId, FormTypes, Id } from './types/public.types';
-export type { Configs } from './types/configs.type';
+export type { Configs } from '../../../types/configs.type';
 
 export default FormBuilder;
