+++ packages/core/src/modules/form/src/utils/selector/formSelector.types.ts
@@ -1,3 +1,3 @@
 import { FormBuilderProps } from '../../components/builder/formBuilder.types';
 
-export type SelectorProps = FormBuilderProps;
+export type FormSelectorProps = FormBuilderProps;
