+++ packages/core/src/types/configs.type.ts
